% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_multiple_variogram.R
\name{fit_multiple_variogram}
\alias{fit_multiple_variogram}
\title{Fitting Multiple Variograms}
\usage{
fit_multiple_variogram(data, coords = c("X", "Y"), length = 99)
}
\arguments{
\item{data}{input data set one of sp, sf or data.frame with X and Y as variables}

\item{coords}{(vector) pair of coordinates if data type is aspatial}

\item{length}{(integer) the length of sequence used to produced fitted data.}
}
\value{
A data frame with combined information from all variables as response is returned.
The \code{data.frame} contains the experimental variogram information, the fitted
variogram model information.
The sums of squares (sserr) between the sample variogram dropped,
For further information please refer to \code{\link[automap]{autofitVariogram}}
or \code{\link{fit_variogram}}
}
\description{
This function returns experimental and fitted variogram information
for multiple variables. For each variable \code{\link[automap]{autofitVariogram}} is used.
The function supports spatial data in simple features (sf), SpatialPointsDataFrame(sp)
or data.frame. If the input data is in \code{data.frame}, the X, and Y should be
present in the data as variables.
}
\examples{

## Read data
data(landcover)
# subset data

lcdat <- landcover[, c("MBLU", "MNDVI", "MNIR", "MPC1", "MPC2", "MPC23", "MRED", "MSAVI")]

# fit variograms

fmv <- fit_multiple_variogram(data = lcdat,coords = NULL,length = 99 )

\dontrun{
data("landcover")
fit_multiple_variogram(data = landcover,coords = NULL, length = 99)
}

}
\references{
\href{https://CRAN.R-project.org/package=automap}{automatp}: A index page for automap package on \href{https://cran.r-project.org}{CRAN}

Hiemstra, P.H., Pebesma, E.J., Twenhofel, C.J.W. and G.B.M. Heuvelink, 2008. Real-time automatic interpolation of ambient gamma dose rates from
the Dutch Radioactivity Monitoring Network. Computers & Geosciences.\href{http://dx.doi.org/10.1016/j.cageo.2008.10.011}{DOI:}
}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_variogram}},
\code{\link{plot_variogram}}
}
