% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_variogram.R
\name{fit_variogram}
\alias{fit_variogram}
\title{Variogram Fitting on Quantitative Data}
\usage{
fit_variogram(data, response, coords = c("X", "Y"))
}
\arguments{
\item{data}{An object of \code{\link[sp]{SpatialPointsDataFrame-class}},\code{\link{sf}} or \code{\link{data.frame}}}

\item{response}{A response variable within quotes e.g. "response"}

\item{coords}{(vector) A pair of coordinate when input data is data.frame and not "sf" or "sp" object. The X (Longitude) and Y (Latitude) should be column variables
in the input dataset}
}
\value{
An object of type autofitVariogram is returned. This object contains the experimental variogram, the fitted variogram model and the sums of squares (sserr) between the sample variogram and the fitted variogram model.
}
\description{
This function is a wrapper around \code{\link[automap]{autofitVariogram}}. The function only runs a simple and ordinary
 krigging. For example, if the respose variable is `R` then the formula that will be supplied in the data is `R~1`.
   Five experimental variogram models are tested automatically on selected response variable. The function support spatial data in simple features (sf),
   SpatialPointsDataFrame(sf) or data.frame. The later format requires X, and Y as variables in the data.frame.
}
\details{
Five different types of geostatistical models are tested through \code{\link[automap]{autofitVariogram}}, they are:
 Spherical (Sph), Gaussian (Gau), Exponential (Exp), Matern family (Mat), and M.Stein's parameterization on Matern model.
 for more information please refer to original \code{\link[automap]{autofitVariogram}}.
}
\examples{

\dontrun{
data("landcover")

vfit<- fit_variogram(data = landcover,response = "ASYM",coords = NULL)
vfit
}

}
\references{
\href{https://CRAN.R-project.org/package=automap}{automatp}: A index page for automap package on cran
}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{multiple_variogram}},
\code{\link{plot_variogram}}
}
