% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_class.R
\name{generate_random_class}
\alias{generate_random_class}
\title{A function to create random multiclass data.
This function allows you to create random spatial data for a specified number of classes (num_class)
and replicates (num_rep)}
\usage{
generate_random_class(data, num_class, num_rep, return_sf = TRUE)
}
\arguments{
\item{data}{Input extent of data 'sf', 'sp', or 'Vect'}

\item{num_class}{number of required classes}

\item{num_rep}{replication for each class}

\item{return_sf}{(logical) if true returns 'sf' object using coordinate
system of input data}
}
\value{
If 'return_sf' is TRUE, the function returns an 'sf' object containing the generated classes.
If 'return_sf' is FALSE, the function returns a data frame containing the generated classes.
}
\description{
A function to create random multiclass data.
This function allows you to create random spatial data for a specified number of classes (num_class)
and replicates (num_rep)
}
\examples{
# Load sample data
data("landcover")

# Generate random classes with 5 classes and 100 replicates per class
samp_data <- generate_random_class(landcover, 5, 100,return_sf = FALSE)
head(samp_data)

}
