% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_variogram.R
\name{multiple_variogram}
\alias{multiple_variogram}
\title{Multiple Variogram Fitting}
\usage{
multiple_variogram(data, coords = c("X", "Y"), length = 99)
}
\arguments{
\item{data}{input data set one of sp, sf or data.frame with X and Y as variables}

\item{coords}{(vector) pair of coordinates if data type is aspatial}

\item{length}{(integer) the length of sequence used to produced fitted data.}
}
\value{
A data frame with combined information from all variables as response is returned.
The dataframe contains the experimental variogram information, the fitted variogram model information.
The sums of squares (sserr) between the sample variogram and the fitted variogram model is dropped, please refer to \code{\link[automap]{autofitVariogram}} or \code{\link{fit_variogram}}
}
\description{
This function returns experimental and fitted variogram information for multiple variables. For each variable \code{\link[automap]{autofitVariogram}} is used.
}
\examples{

\dontrun{
data("landcover")
multiple_variogram(data = landcover,coords = NULL, length = 99)
}

}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_variogram}},
\code{\link{plot_variogram}}
}
