% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_variogram.R
\name{plot_multiple_variogram}
\alias{plot_multiple_variogram}
\title{Plotting Multiple Variograms}
\usage{
plot_multiple_variogram(object, plotType = "line", showRange = TRUE)
}
\arguments{
\item{object}{An output from \code{\link{multiple_variogram}} function}

\item{plotType}{plot type "bar" or "line". "linePoint" or "pointLine" or "line", or "point" are used to use
point and line format.}

\item{showRange}{(logical) whether to display range in the plot}
}
\value{
A list with multiple semivariance plot for all variables and a tibble with fitted semivariances for all variables.
}
\description{
This function uses ggplot2 framework to plot semivariogram object (`tibble`) result from
\code{\link{multiple_variogram}}.
}
\examples{
\dontrun{
data("landcover")
landcover$Class_name<- NULL
mfit  <- multiple_variogram(data = landcover,coords = NULL, length = 99)
mplot <- plot_multiple_variogram (mfit, plotType = "line", showRange = FALSE)

# Display the plot
mplot$plot

# View tibble
mplot$data
}


}
\references{
H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_variogram}},
\code{\link{plot_variogram}}
}
