% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variogram.R
\name{plot_variogram}
\alias{plot_variogram}
\title{Plotting variogram using ggplot2}
\usage{
plot_variogram(object, length = 99, showRange = TRUE)
}
\arguments{
\item{object}{An output from \code{\link{fit_variogram}} function}

\item{length}{(integer) the length of sequence used to produced fitted data.}

\item{showRange}{(logical) whether to display range in the plot}
}
\value{
A list with plot and tibble with semivariogram information
}
\description{
This function uses ggplot2 framework to plot semivariogram object results from
\code{\link{fit_variogram}}.
}
\examples{

\dontrun{
data("landcover")

vfit<- fit_variogram(data = landcover,response = "ASYM",coords = NULL)
vplot<- plot_variogram(object = vfit, length = 99, showRange = TRUE)
# plot variogram
vplot$plot

}


}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_variogram}},
\code{\link{plot_variogram}}
}
