% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_features.R
\name{rotate_features}
\alias{rotate_features}
\title{Rotate Features}
\usage{
rotate_features(input_sf, rotation_angle = 45, rotation_base = "center")
}
\arguments{
\item{input_sf}{The input spatial object to be rotated.
It can be an 'sf' object or a 'Spatial' object.}

\item{rotation_angle}{The angle of rotation in degrees.}

\item{rotation_base}{The rotation base. It can be "center" or a
coordinate specifying the rotation base.}
}
\value{
A list containing the rotation base and the rotated features as an 'sf' object.
}
\description{
This function rotates the features of an input spatial object by the
specified rotation angle and rotation base.
}
\examples{
# Load sample data
data(landcover)

# Rotate the landcover data by 45 degrees around the center

rot_coords <- rotate_features(input_sf = landcover, rotation_angle = 45, rotation_base = "center")

}
