% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_cluster_sample.R
\name{spatial_cluster_sample}
\alias{spatial_cluster_sample}
\title{Spatial cluster  cross-validation}
\usage{
spatial_cluster_sample(
  data = data,
  coords = NULL,
  v = 10,
  spatial = TRUE,
  clust_method = "kmeans",
  dist_clust = NULL,
  ...
)
}
\arguments{
\item{data}{input data set one of sp, sf or data.frame with X and Y as variables}

\item{coords}{(vector) pair of coordinates if data type is aspatial}

\item{v}{number of partitions of the dataset or number of clusters}

\item{spatial}{(logical) if data set is spatial (when sf or sp) or aspatial (data.frame)}

\item{clust_method}{one of partitioning (default = kmeans) or one of hierarchical methods(\code{\link[stats]{hclust}})}

\item{dist_clust}{the agglomeration method to be used. This should be one of “ward.D”, “ward.D2”, “single”, “complete”,
“average” (= UPGMA), “mcquitty” (= WPGMA), “median” (= WPGMC) or “centroid” (= UPGMC).
the dist_clust in the funtion is method in stats::hclust}

\item{...}{Extra arguments passed on to [stats::kmeans()] or [stats::hclust()]}
}
\value{
A tibble with classes `spatial_cv`, `rset`, `tbl_df`, `tbl`, and
 `data.frame`. The results include a column for the data split objects and
 an identification variable `id`.
}
\description{
This is a wrapper functin around \code{\link[spatialsample]{spatial_clustering_cv}}.
 Spatial cluster sampling splits the data into V groups groups using partitioning (kmeans)/ hierarchical(hclust) clustering of some variables, typically
 spatial coordinates. A resample of the analysis data consists of V-1 of the folds/clusters while the assessment set contains the final fold/cluster. In
 basic spatial cross-validation (i.e. no repeats), the number of resamples.
 is equal to V.
}
\details{
The variables in the `coords` argument, if input data is data.frame or extracted from sp, or sf data are used for  clustering of
 the data into disjointed sets. These clusters are used as the folds for cross-validation. Depending on how the
 data are distributed spatially.
}
\references{
A. Brenning, "Spatial cross-validation and bootstrap for the assessment of
prediction rules in remote sensing: The R package sperrorest," 2012 IEEE
International Geoscience and Remote Sensing Symposium, Munich, 2012,
pp. 5372-5375, doi: 10.1109/IGARSS.2012.6352393.

\href{https://spatialsample.tidymodels.org/}{spatialsample}: provide functions and classes for spatial resampling to use with rsample
}
