% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_hexgrid_sample.R
\name{spatial_hexgrid_sample}
\alias{spatial_hexgrid_sample}
\title{Spatial Hexagonal Grid Sampling}
\usage{
spatial_hexgrid_sample(
  data,
  sidelength,
  top_type = "flat",
  show_grid = TRUE,
  fold_selection = "default",
  k = NULL
)
}
\arguments{
\item{data}{The input data. It can be either an \code{sf} object or a \code{SpatialPoints} object.}

\item{sidelength}{The length of the hexagon's side.}

\item{top_type}{The type of the hexagon top, either "flat" or "pointy".
Default is "flat".}

\item{show_grid}{Logical value indicating whether to display the hexagonal grid.
Default is TRUE.}

\item{fold_selection}{The method for fold selection. Options include "default",
"systematic", or "random". Default is "default".}

\item{k}{The number of folds for fold selection. Required when
'fold_selection' is "systematic" or "random". Default is NULL.}
}
\value{
A list containing the plots (if 'show_grid' is TRUE),
the hexagonal blocks (sf object), and the split objects.
}
\description{
This function generates a hexagonal grid and performs spatial sampling based
on the provided data and parameters.
}
\examples{
# Load sample data
data(landcover)

# Generate a spatial hexagonal grid sampling with a sidelength of
# 10 km and 5 folds using systematic selection
hexgrid <- spatial_hexgrid_sample(landcover, sidelength = 10000,
            fold_selection = "systematic", k = 5)


}
