% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_quadgrid_sample.R
\name{spatial_quadgrid_sample}
\alias{spatial_quadgrid_sample}
\title{Spatial Quadgrid Sampling}
\usage{
spatial_quadgrid_sample(
  data,
  cellsize = c(10000, 10000),
  offset = c(0, 0),
  show_grid = TRUE,
  rotation_angle = NULL,
  fold_selection = "default",
  k = NULL
)
}
\arguments{
\item{data}{The input data. It can be either an 'sf' object or a \code{SpatialPoints} object.}

\item{cellsize}{The size of the grid cells, provided as a numeric vector
with two values representing the x and y dimensions.}

\item{offset}{The offset of the grid, provided as a numeric vector with
two values representing the x and y offsets.}

\item{show_grid}{Logical value indicating whether to display the \code{quadgrid}.
Default is TRUE.}

\item{rotation_angle}{The angle of rotation for the \code{quadgrid}. Default is NULL.}

\item{fold_selection}{The method for fold selection. Options include
"default", "systematic", or "random". Default is "default".}

\item{k}{The number of folds for fold selection. Required when
\code{fold_selection} is "systematic" or "random". Default is NULL.}
}
\value{
A list containing the plots (if 'show_grid' is TRUE),
the \code{quadgrid} blocks (sf object), and the split objects.
}
\description{
This function generates a regular rectangular grids and performs spatial
sampling based on the provided data and parameters.
The result from this function is similar to \verb{Generate Tessellation}
tool in ESRI's ArcGIS software but only square or rectangle polygons are
possible. The extent of the point coordinates is
divided into number of possible grids based on the values of \code{cellsize}.
The \code{cellsize} is the length and width of polygon to be created.
At this point \code{spatial_quadgrid_sample} and \code{spatial_hexgrid_sample}are different
function. The reason is because rotation of original data and resultant grid
is possible for regular rectangular grids. In the future both functions can be merged
together
}
\details{
The spatial grid sample generates tessellation or fishnet in ESRI's ArcGIS software,
Currently \code{sp}, \code{sf} and \code{Vect} data are allowed.
The grid shows the number of samples within the grid and grid number
e.g. (80,1). The \code{default} selection  numbers the grids from bottom left
and increases row wise.
The \code{random} selection assigns the grid number randomly.
The \code{systematic} grid numbering happens from bottom right and column wise.

The \code{random} selection as its' name applies select grids randomly, and
\code{systematic} selection option
allows to select the grids sequentially up to \code{k}
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for Spatial Vector Data. The R
Journal 10 (1), 439-446, https://doi.org/10.32614/RJ-2018-009
}
