% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_thinning_class.R
\name{spatial_thinning_class}
\alias{spatial_thinning_class}
\title{Spatial Thinning of Multi-class Point Data}
\usage{
spatial_thinning_class(
  data,
  spatial = FALSE,
  coords = c("X", "Y"),
  distance,
  reps = 1,
  class = "class"
)
}
\arguments{
\item{data}{input data set one of sp, sf or data.frame with X and Y as variables}

\item{spatial}{(logical) if variable is spatial or aspatial}

\item{coords}{(vector) pair of coordinates if data type is aspatial}

\item{distance}{minimum distance between two points of same class}

\item{reps}{(integer) replication}

\item{class}{multi-class variable or strata variable. For each class
thinning will be applied based on specified distance}
}
\value{
data.frame
}
\description{
The function is a wrapper around spatial thinning \code{\link[spThin]{thin}} function of spThin package.
 It makes spatial thinning easy for multi-class input data set. This function is handy and flexible:
 can support xy coordinates, spatialPoints, or simple features. Unlike spthin, this function assumes that data are in projected
 planar coordinates system; preferably Universal Transverse Mercator (UTM).
 There is no restrictions on number of columns in the data set as long as data is in the required format.
}
\details{
Spatial thinning expect the data in the projected coordinate system, the distance
}
\examples{

\dontrun{
data("landcover")
spatial_thinning_class(data = landcover, spatial = TRUE, coords = NULL,
distance = 10000, reps = 2, class = "Class_namme")
}

}
\seealso{
\code{\link[spThin]{thin}}
}
