% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_variogram.R
\name{fit_variogram}
\alias{fit_variogram}
\title{Variogram Fitting on Quantitative Data}
\usage{
fit_variogram(data, response, coords = c("X", "Y"))
}
\arguments{
\item{data}{An object of \code{\link[sp]{SpatialPointsDataFrame-class}},
\code{\link{sf}} or \code{\link{data.frame}}}

\item{response}{A response variable within quotes e.g. "response"}

\item{coords}{(vector) A pair of coordinate when input data is data.frame
and not "sf" or "sp" object. The X (Longitude) and Y (Latitude) should be
column variables in the input data set}
}
\value{
An object of a type autofitVariogram is returned. This object contains
the experimental variogram,the fitted variogram model and the sums of
squares (sserr) between the sample variogram and the fitted variogram model.
}
\description{
This function is a wrapper around \code{\link[automap]{autofitVariogram}}.
The function runs a simple and ordinary krigging. For example, if the r
esponse variable is \code{R} then the formula that will be supplied in the data is \code{R~1}.

Five experimental variogram models are tested automatically on selected
response variable. The function supports spatial data in simple features
(sf), SpatialPointsDataFrame(sp) or data.frame.
The later format requires X, and Y as variables in the data.frame.
}
\details{
Five different types of geostatistical models are tested
through \code{\link[automap]{autofitVariogram}}, they are:Spherical (Sph),
Gaussian (Gau), Exponential (Exp), Matern family (Mat), and
M.Stein's parameterization on Matern model.
For more information please refer to original \code{\link[automap]{autofitVariogram}} function from automap package.
}
\examples{
\dontrun{
data("landcover")

vfit<- fit_variogram(data = landcover,response = "MPC1",coords = NULL)
vfit
}

}
\references{
Hiemstra, P.H., Pebesma, E.J., Twenhofel, C.J.W. and G.B.M. Heuvelink, 2008.
Real-time automatic interpolation of ambient gamma dose rates from the Dutch
Radioactivity Monitoring Network. Computers & Geosciences.\href{http://dx.doi.org/10.1016/j.cageo.2008.10.011}{DOI:}

Pebesma, E.J., 2004. Multivariable geostatistics in S: the gstat package. Computers & Geosciences, 30: 683-691.
Benedikt Gräler, Edzer Pebesma and Gerard Heuvelink, 2016. Spatio-Temporal Interpolation using gstat.  The R
Journal 8(1), 204-218
}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_multiple_variogram}},
\code{\link{plot_variogram}}
}
