% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_variogram.R
\name{multiple_variogram}
\alias{multiple_variogram}
\title{Multiple Variogram Fitting}
\usage{
multiple_variogram(data, coords = c("X", "Y"), length = 99)
}
\arguments{
\item{data}{input data set one of sp, sf or data.frame with X and Y as variables}

\item{coords}{(vector) pair of coordinates if data type is aspatial}

\item{length}{(integer) the length of sequence used to produced fitted data.}
}
\value{
A data frame with combined information from all variables as response is returned.
The `data.frame` contains the experimental variogram information, the fitted variogram model information.
The sums of squares (sserr) between the sample variogram dropped,
For further information please refer to \code{\link[automap]{autofitVariogram}} or \code{\link{fit_variogram}}
}
\description{
This function returns experimental and fitted variogram information for multiple variables.
For each variable \code{\link[automap]{autofitVariogram}} is used. The function supports
   spatial data in simple features (sf), SpatialPointsDataFrame(sp) or data.frame.
   If the input data is in `data.frame`, the X, and Y should be present in the data as variables.
}
\examples{

## Read data
lcdat<- landcover

# remove factor variable

# lcdat$class_name <- NULL

# fit variograms

fmv<- multiple_variogram(data = lcdat,coords = NULL,length = 99 )

\dontrun{
data("landcover")
fit_multiple_variogram(data = landcover,coords = NULL, length = 99)
}

}
\references{
\href{https://CRAN.R-project.org/package=automap}{automatp}: A index page for automap package on [CRAN](https://cran.r-project.org)

 Hiemstra, P.H., Pebesma, E.J., Twenhofel, C.J.W. and G.B.M. Heuvelink, 2008. Real-time automatic interpolation of ambient gamma dose rates from
 the Dutch Radioactivity Monitoring Network. Computers & Geosciences.[DOI:](http://dx.doi.org/10.1016/j.cageo.2008.10.011)
}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_variogram}},
\code{\link{plot_variogram}}
}
