% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_variogram.R
\name{plot_multiple_variogram}
\alias{plot_multiple_variogram}
\title{Plotting multiple variograms using ggplot2}
\usage{
plot_multiple_variogram(object, plotType = "line", showRange = TRUE)
}
\arguments{
\item{object}{An output from \code{\link{multiple_variogram}} function}

\item{plotType}{plot type "bar" or "line". "linePoint" or "pointLine" or "line", or "point" are used to use
point and line format.}

\item{showRange}{(logical) whether to display range in the plot}
}
\value{
A list with multiple semivariance plot and tibble with semivariogram information
}
\description{
This function uses ggplot2 framework to plot semivariogram object result from
\code{\link{multiple_variogram}}.
}
\examples{
\dontrun{
data("landcover")
mfit  <- multiple_variogram(data = landcover,coords = NULL, length = 99)
mplot <- plot_multiple_variogram

# plot information
mplot$plot
# table
mplot$data
}


}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_variogram}},
\code{\link{plot_variogram}}
}
