% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_variogram.R
\name{plot_multiple_variogram}
\alias{plot_multiple_variogram}
\title{Plotting Multiple Variograms}
\usage{
plot_multiple_variogram(object, plot_type = "line", show_range = TRUE)
}
\arguments{
\item{object}{An output from \code{\link{fit_multiple_variogram}} function}

\item{plot_type}{plot type "bar" or "line". "linePoint" or "pointLine" or
"line", or "point" are used to use
point and line format.}

\item{show_range}{(logical) whether to display range in the plot}
}
\value{
A list with multiple semivariance plot for all variables and a tibble
with fitted semivariances for all variables.
}
\description{
This function uses ggplot2 framework to plot semivariogram object (\code{tibble})
result from \code{\link{fit_multiple_variogram}}.
}
\examples{
##  Run semivariance on Principal Component Axis1
# read data
data("landcover")

# remove factor variable

landcover$class_name <- NULL

# fit multiple variograms

mvfit  <- fit_variogram(data = landcover, response = "MPC1", coords = NULL)

mvplot <- plot_variogram(object = mvfit, length = 99, show_range = TRUE)


\dontrun{
# load data
data("landcover")

# remove factor variable
landcover$Class_name<- NULL

# fit multiple variograms
mfit  <- multiple_variogram(data = landcover,coords = NULL, length = 99)

mplot <- plot_multiple_variogram (mfit, plot_type = "line", show_range = FALSE)

# Display the plot

mplot$plot

# View tibble
mplot$data
}
}
\references{
Pebesma, E.J., 2004. Multivariable geostatistics in S: the gstat package. Computers & Geosciences, 30: 683-691.
Benedikt Gräler, Edzer Pebesma and Gerard Heuvelink, 2016. Spatio-Temporal Interpolation using gstat.  The R
Journal 8(1), 204-218

Hiemstra, P.H., Pebesma, E.J., Twenhofel, C.J.W. and G.B.M. Heuvelink, 2008. Real-time automatic interpolation of ambient gamma dose rates from
the Dutch Radioactivity Monitoring Network. Computers & Geosciences.\href{http://dx.doi.org/10.1016/j.cageo.2008.10.011}{DOI:}

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_variogram}},
\code{\link{plot_variogram}}
}
