% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variogram.R
\name{plot_variogram}
\alias{plot_variogram}
\title{Plot Single Sevariance Model}
\usage{
plot_variogram(object, length = 99, show_range = TRUE)
}
\arguments{
\item{object}{An output from \code{\link{fit_variogram}} function}

\item{length}{(integer) the length of sequence used to produced fitted data.}

\item{show_range}{(logical) whether to display range in the plot}
}
\value{
A list with plot and tibble with semivariogram information
}
\description{
This function uses ggplot2 framework to plot semivariogram object results from
\code{\link{fit_variogram}}.
}
\examples{

##  Run semivariance on Principal Component Axis1
# read data
data("landcover")

vfit  <- fit_variogram(data = landcover, response = "MPC1", coords = NULL)
vplot <- plot_variogram(object = vfit, length = 99, show_range = TRUE)

# plot
 vplot$plot

vfit_pc2  <- fit_variogram(data = landcover, response = "MPC1", coords = NULL)
vplot_pc2  <- plot_variogram(object = vfit_pc2, length = 99, show_range = FALSE)

# plot
 vplot_pc2$plot

\dontrun{
data("landcover")

vfit  <- fit_variogram(data = landcover,response = "ASYM",coords = NULL)
vplot <- plot_variogram(object = vfit, length = 99, show_range = TRUE)

# plot variogram
 vplot$plot
# data
vplot$data

}


}
\references{
Pebesma, E.J., 2004. Multivariable geostatistics in S: the gstat package. Computers & Geosciences, 30: 683-691.
Benedikt Gräler, Edzer Pebesma and Gerard Heuvelink, 2016. Spatio-Temporal Interpolation using gstat.  The R
Journal 8(1), 204-218

Hiemstra, P.H., Pebesma, E.J., Twenhofel, C.J.W. and G.B.M. Heuvelink, 2008. Real-time automatic interpolation of ambient gamma dose rates from
the Dutch Radioactivity Monitoring Network. Computers & Geosciences.\href{http://dx.doi.org/10.1016/j.cageo.2008.10.011}{DOI:}

H. Wickham. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.
}
\seealso{
\code{\link[automap]{autofitVariogram}}, \code{\link{fit_variogram}},
\code{\link{plot_variogram}}
}
