% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_thinning_class.R
\name{spatial_thinning_class}
\alias{spatial_thinning_class}
\title{Spatial Thinning of Multi-class Point Data}
\usage{
spatial_thinning_class(
  data,
  spatial = FALSE,
  coords = c("X", "Y"),
  distance,
  reps = 1,
  class = "class"
)
}
\arguments{
\item{data}{input data set one of sp, sf or data.frame with X and Y as variables}

\item{spatial}{(logical) if variable is spatial or aspatial}

\item{coords}{(vector) pair of coordinates if data type is aspatial}

\item{distance}{minimum distance between two points of same class}

\item{reps}{(integer) replication}

\item{class}{multi-class variable or strata variable. For each class
thinning will be applied based on specified distance}
}
\value{
data.frame
}
\description{
The function is a wrapper around spatial thinning \code{\link[spThin]{thin}} function of spThin package.
 It makes spatial thinning easy for multi-class input data set. This function is handy and flexible:
 can support xy coordinates, spatialPoints, or simple features. Unlike spthin, this function assumes that data are in projected
 planar coordinates system; preferably Universal Transverse Mercator (UTM).
 There is no restrictions on number of columns in the data set as long as data is in the required format.
}
\details{
Spatial thinning expect the data in the projected coordinate system, the distance
 value is the threshold which limits the number of observations. All the observations, in a particular class, smaller than the
 distance thresholds are deleted. If there are more than two classes, minimum distance between observations from each class.

 @references
Aiello-Lammens, M. E., Boria, R. A., Radosavljevic, A. , Vilela, B. and Anderson, R. P. (2015). spThin: an
Rpackage for spatial thinning of species occurrence records for use in ecological niche models.
Ecography, 38:541-545. URL https://onlinelibrary.wiley.com/doi/10.1111/ecog.01132.
}
\examples{
## load data from the package
tdat<- landcover

# Run spatial thinning (sf format)
stc<- spatial_thinning_class(data = tdat,spatial = TRUE,
coords = NULL,distance = 10000,reps = 1,class = "Class_name")

### compare the number of samples from original data to the thinned data
## Original data
library(dplyr)
tdat \%>\% group_by(Class_name) \%>\% summarize(n = n())

## Thinned data
stc \%>\% group_by(Class_name) \%>\% summarize(n = n())

##  spatial thinning using meuse data
# load library and data
library(sp)
data(meuse)

# apply thinning

mthin<- spatial_thinning_class(data = meuse, spatial = FALSE,
         coords = c("x", "y"),
         distance = 1000,
         reps = 2, class = "ffreq"
         )

\dontrun{
data<- landcover
spatial_thinning_class(data = landcover, spatial = TRUE, coords = NULL,
distance = 10000, reps = 2, class = "Class_name")
}

}
\seealso{
\code{\link[spThin]{thin}}
}
