% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdcab-package.R
\docType{package}
\name{stdcab-package}
\alias{stdcab}
\alias{stdcab-package}
\title{stdcab: Spatial Thinning, Dependency, Clustering, and Blocking of Spatial Point Data for Classification Problems}
\description{
Functions to spatial point data for spatial thinning, Cluster analysis for random cross-validation and repeated-cross validation samples creation, and spatial block cross validation. Spatial structure in the data accounted for using variogram and spatial blocking of observations. Results from spatial clustering and blocking are ready to be used in the tidy-modeling framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/suvedimukti/stdcab}
  \item Report bugs at \url{https://github.com/suvedimukti/stdcab/issues}
}

}
\author{
\strong{Maintainer}: Mukti Subedi \email{suvedimukti@gmail.com}

}
\keyword{internal}
