% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdcab-package.R
\docType{package}
\name{stdcab-package}
\alias{stdcab}
\alias{stdcab-package}
\title{stdcab: Spatial Thinning, Dependency, Clustering, And Blocking of Point Data for Classification Problems}
\description{
Functions to spatial point data for spatial thinning, clustering, blocking and structure analysis. Results from spatial clustering and blocking are resampling objects for machine learning using tidy modelling framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/suvedimukti/stdcab}
  \item \url{https://suvedimukti.github.io/stdcab/}
  \item Report bugs at \url{https://github.com/suvedimukti/stdcab/issues}
}

}
\author{
\strong{Maintainer}: Mukti Subedi \email{suvedimukti@gmail.com} (\href{https://orcid.org/0000-0003-4167-0206}{ORCID})

}
\keyword{internal}
